+++ pkg/shares/parse_contiguous_shares.go
